refactor(Core/sWorld): improve singleton management (#3862)
